View working.

+++ forecast/models.py
@@ -616,6 +616,7 @@ class ForecastBudgetDataView(models.Model):
     """Used for joining budgets and forecast.
     Mapped to a view in the database, because
     the query is too complex"""
+    id = models.IntegerField(primary_key=True, )
 
     # The view is created by  a migration
     financial_code = models.ForeignKey(
@@ -623,19 +624,19 @@ class ForecastBudgetDataView(models.Model):
         on_delete=models.PROTECT,
     )
     financial_year = models.IntegerField()
-    Budget = models.BigIntegerField(default=0)
-    Apr = models.BigIntegerField(default=0)
-    May = models.BigIntegerField(default=0)
-    Jun = models.BigIntegerField(default=0)
-    Jul = models.BigIntegerField(default=0)
-    Aug = models.BigIntegerField(default=0)
-    Sep = models.BigIntegerField(default=0)
-    Oct = models.BigIntegerField(default=0)
-    Nov = models.BigIntegerField(default=0)
-    Dec = models.BigIntegerField(default=0)
-    Jan = models.BigIntegerField(default=0)
-    Feb = models.BigIntegerField(default=0)
-    Mar = models.BigIntegerField(default=0)
+    budget = models.BigIntegerField(default=0)
+    apr = models.BigIntegerField(default=0)
+    may = models.BigIntegerField(default=0)
+    jun = models.BigIntegerField(default=0)
+    jul = models.BigIntegerField(default=0)
+    aug = models.BigIntegerField(default=0)
+    sep = models.BigIntegerField(default=0)
+    oct = models.BigIntegerField(default=0)
+    nov = models.BigIntegerField(default=0)
+    dec = models.BigIntegerField(default=0)
+    jan = models.BigIntegerField(default=0)
+    feb = models.BigIntegerField(default=0)
+    mar = models.BigIntegerField(default=0)
 
     class Meta:
         managed = False
@@ -684,32 +685,31 @@ class OSCARReturn(models.Model):
 
 
 """
-DROP VIEW
+DROP VIEW if exists
 forecast_forecast_budget_view ;
 
 CREATE VIEW
 forecast_forecast_budget_view 
 as
-SELECT forecast_forecastmonthlyfigure.financial_code_id, "forecast_forecastmonthlyfigure"."financial_year_id" as financial_year,
-SUM("forecast_budgetmonthlyfigure"."amount") AS "Budget",
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Apr' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Apr", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Aug' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Aug", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Dec' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Dec", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Feb' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Feb", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jan' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jan", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jul' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jul", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jun' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jun", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Mar' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Mar", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'May' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "May", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Nov' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Nov", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Oct' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Oct", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Sep' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Sep" 
+SELECT SELECT row_number() OVER () as id, forecast_forecastmonthlyfigure.financial_code_id, "forecast_forecastmonthlyfigure"."financial_year_id" as financial_year,
+COALESCE (SUM("forecast_budgetmonthlyfigure"."amount"),0) AS "budget",
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 1 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "apr", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 2 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "may", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 3 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "jun", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 4 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "jul", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 5 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "aug", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 6 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "sep", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 7 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "oct", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 8 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "nov", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 9 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "dec", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 10 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "jan", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 11 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "feb", 
+       COALESCE(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 12 THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END), 0) AS "mar" 
 FROM "forecast_forecastmonthlyfigure"
     INNER JOIN "forecast_financialperiod" ON ("forecast_forecastmonthlyfigure"."financial_period_id" = "forecast_financialperiod"."financial_period_code")
 	LEFT OUTER  JOIN forecast_budgetmonthlyfigure ON 
 	forecast_forecastmonthlyfigure.financial_code_id = "forecast_budgetmonthlyfigure"."financial_code_id"
 	AND forecast_forecastmonthlyfigure.financial_year_id = "forecast_budgetmonthlyfigure"."financial_year_id"
-WHERE "forecast_forecastmonthlyfigure"."financial_year_id" = 2019 
 GROUP BY forecast_forecastmonthlyfigure.financial_code_id, "forecast_forecastmonthlyfigure"."financial_year_id"
 
 
@@ -720,48 +720,9 @@ GROUP BY forecast_forecastmonthlyfigure.financial_code_id, "forecast_forecastmon
 
 
 
-"""SELECT financial_code_id,        
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Apr' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Apr", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Aug' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Aug", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Dec' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Dec", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Feb' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Feb", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jan' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jan", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jul' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jul", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jun' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jun", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Mar' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Mar", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'May' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "May", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Nov' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Nov", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Oct' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Oct", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Sep' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Sep" 
-FROM "forecast_forecastmonthlyfigure"
-    INNER JOIN "forecast_financialperiod" ON ("forecast_forecastmonthlyfigure"."financial_period_id" = "forecast_financialperiod"."financial_period_code")
-WHERE "forecast_forecastmonthlyfigure"."financial_year_id" = 2019 
-GROUP BY financial_code_id
-"""
 
 
-"""
-SELECT forecast_forecastmonthlyfigure.financial_code_id, 
-SUM("forecast_budgetmonthlyfigure"."amount") AS "budget",
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Apr' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Apr", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Aug' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Aug", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Dec' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Dec", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Feb' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Feb", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jan' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jan", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jul' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jul", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jun' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Jun", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Mar' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Mar", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'May' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "May", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Nov' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Nov", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Oct' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Oct", 
-       SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Sep' THEN "forecast_forecastmonthlyfigure"."amount" ELSE NULL END) AS "Sep" 
-FROM "forecast_forecastmonthlyfigure"
-    INNER JOIN "forecast_financialperiod" ON ("forecast_forecastmonthlyfigure"."financial_period_id" = "forecast_financialperiod"."financial_period_code")
-	LEFT OUTER  JOIN forecast_budgetmonthlyfigure ON forecast_forecastmonthlyfigure.financial_code_id = "forecast_budgetmonthlyfigure"."financial_code_id"
-WHERE "forecast_forecastmonthlyfigure"."financial_year_id" = 2019 
-GROUP BY forecast_forecastmonthlyfigure.financial_code_id
 
-"""
 
 
 """
@@ -773,18 +734,18 @@ CREATE VIEW "forecast_oscarreturn" as
             account_l5_code,
             "treasurySS_subsegment"."sub_segment_code" ,
             "treasurySS_subsegment"."sub_segment_long_name" ,                    
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Apr' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "apr",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Aug' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "aug",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Dec' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "dec",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Feb' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "feb",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jan' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "jan",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jul' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "jul",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Jun' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "jun",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Mar' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "mar",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'May' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "may",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Nov' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "nov",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Oct' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "oct",
-            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."period_short_name" = 'Sep' THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "sep"
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 1 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "apr",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 2 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "may",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 3 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "jun",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 4 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "jul",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 5 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "aug",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 6 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "sep",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 7 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "oct",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 8 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "nov",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 9 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "dec",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 10 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "jan",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 11 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "feb",
+            coalesce(round(SUM(CASE WHEN "forecast_financialperiod"."financial_period_code" = 12 THEN "forecast_monthlyfigureamount"."amount" ELSE NULL END)/100000), 0)  AS "mar"
         FROM "forecast_monthlyfigureamount"
                 INNER JOIN
                  "forecast_monthlyfigure" on (forecast_monthlyfigureamount.monthly_figure_id = forecast_monthlyfigure.id)
